data: Move the XML installation dir to /usr/share/xml
This is more standard (e.g. D-Bus does it, installing its files in
`/usr/share/xml/dbus-1`), and gives us a nice place to publicly install
the DTD catalog (see the previous commit).

This does mean that projects which looked up the DTD in the old location
will now fail to find it.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>