data: Drop EggDBus annotations from interface XML files
Since commit d04cd425c6049da749516f37252264f96109e179 EggDBus has not
been used to build the documentation for the D-Bus interfaces (it’s
maintained manually in `docs/polkit/`). While it would be good to keep
documentation inline in the interface definitions, this use of
`<annotation>` elements is not valid according to the DTD —
`<annotation>` elements cannot be nested (and their names are supposed
to be unique).

So, delete the old EggDBus annotations.

Another approach to this problem would have been to port the
documentation from these annotations into XML comments, so that it
remains in the XML files, but that wouldn’t fix divergence from the
manually maintained docbook files. That needs some more thinking about
in the long term.

This fixes the `validate-dbus-interfaces` test.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>