No use for authority in polkit_backend_authority_log
The authority argument is actually only used for verifying it's really
a polkit authority and nothing else. This limits the usage of the
function only to parts of the *backend* where authority is passed and
nowhere else. This leads to the need to use g_print or g_error fns to be
used where *authority doesn't exist. Logs passed to glib sometimes do
not reach journal and do not conform to loglevels.